consolidate sync to moesif
